requirements.txt
Django==1.9.5
psycopg2==2.6.1
raven==5.13.0
Jinja2==2.8
environs==1.1.0
django-environ==0.4.0
